+++ warden-roboto/src/main/kotlin/exceptions/Throwables.kt
@@ -1,12 +1,13 @@
 package at.asitplus.attestation.android.exceptions
 
 import at.asitplus.attestation.android.SoftwareAttestationChecker
+import java.security.cert.X509Certificate
 
 /**
  * Base class for all well-defined Android attestation exceptions.
  * If this one is thrown, a well-defined error arose.
  */
-abstract class AndroidAttestationException(message: String?, cause: Throwable?):Throwable(message, cause) {
+abstract class AndroidAttestationException(message: String?, cause: Throwable?) : Throwable(message, cause) {
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (other !is AndroidAttestationException) return false
@@ -26,7 +27,10 @@ abstract class AndroidAttestationException(message: String?, cause: Throwable?):
  * @param cause the underlying exception
  * @param reason one of a set of well-defined [Reason]s why the attestation failed
  */
-class AttestationValueException(message: String?, cause: Throwable? = null, val reason: Reason) : AndroidAttestationException(message, cause) {
+class AttestationValueException(
+    message: String?, cause: Throwable? = null, val reason: Reason,
+    val expectedValue: Any, val actualValue: Any?
+) : AndroidAttestationException(message, cause) {
 
     /**
      * Possible reasons an [AttestationValueException] was thrown
@@ -96,6 +100,8 @@ class AttestationValueException(message: String?, cause: Throwable? = null, val 
         if (!super.equals(other)) return false
 
         if (reason != other.reason) return false
+        if (expectedValue != other.expectedValue) return false
+        if (actualValue != other.actualValue) return false
 
         return true
     }
@@ -103,8 +109,11 @@ class AttestationValueException(message: String?, cause: Throwable? = null, val 
     override fun hashCode(): Int {
         var result = super.hashCode()
         result = 31 * result + reason.hashCode()
+        result = 31 * result + expectedValue.hashCode()
+        result = 31 * result + (actualValue?.hashCode() ?: 0)
         return result
     }
+
 }
 
 /**
@@ -114,7 +123,12 @@ class AttestationValueException(message: String?, cause: Throwable? = null, val 
  * @param cause the underlying exception
  * @param reason one of a set of well-defined [Reason]s why the attestation failed
  */
-class CertificateInvalidException(message: String, cause: Throwable? = null, val reason: Reason) :
+open class CertificateInvalidException(
+    message: String, cause: Throwable? = null,
+    val certificateChain: List<X509Certificate>,
+    val invalidCertificate: X509Certificate?,
+    val reason: Reason
+) :
     AndroidAttestationException(message, cause) {
 
     /**
@@ -138,6 +152,8 @@ class CertificateInvalidException(message: String, cause: Throwable? = null, val
         if (other !is CertificateInvalidException) return false
         if (!super.equals(other)) return false
 
+        if (certificateChain != other.certificateChain) return false
+        if (invalidCertificate != other.invalidCertificate) return false
         if (reason != other.reason) return false
 
         return true
@@ -145,8 +161,72 @@ class CertificateInvalidException(message: String, cause: Throwable? = null, val
 
     override fun hashCode(): Int {
         var result = super.hashCode()
+        result = 31 * result + certificateChain.hashCode()
+        result = 31 * result + (invalidCertificate?.hashCode() ?: 0)
         result = 31 * result + reason.hashCode()
         return result
+    }
+
+
+    class OtherMatchingRoot(
+        message: String,
+        cause: Throwable? = null,
+        certificateChain: List<X509Certificate>,
+        invalidCertificate: X509Certificate?,
+        val rootCertStage: Stage
+    ) : CertificateInvalidException(message, cause, certificateChain, invalidCertificate, reason = Reason.TRUST) {
+        enum class Stage {
+            SOFTWARE,
+            HARDWARE,
+        }
+
+        override fun equals(other: Any?): Boolean {
+            if (this === other) return true
+            if (other !is OtherMatchingRoot) return false
+            if (!super.equals(other)) return false
+
+            if (rootCertStage != other.rootCertStage) return false
+
+            return true
+        }
+
+        override fun hashCode(): Int {
+            var result = super.hashCode()
+            result = 31 * result + rootCertStage.hashCode()
+            return result
+        }
+    }
+
+
+    class NoMatchingRoot(
+        message: String,
+        cause: Throwable? = null,
+        certificateChain: List<X509Certificate>,
+        invalidCertificate: X509Certificate?,
+    ) : CertificateInvalidException(message, cause, certificateChain, invalidCertificate, reason = Reason.TRUST) {
+
+        override fun equals(other: Any?): Boolean {
+            if (this === other) return true
+            if (other !is OtherMatchingRoot) return false
+            if (!super.equals(other)) return false
+            return true
+        }
+    }
+
+    class InvalidRoot(
+        message: String,
+        cause: Throwable? = null,
+        reason: Reason,
+        certificateChain: List<X509Certificate>,
+        invalidCertificate: X509Certificate?,
+    ) : CertificateInvalidException(message, cause, certificateChain, invalidCertificate, reason) {
+
+        override fun equals(other: Any?): Boolean {
+            if (this === other) return true
+            if (other !is OtherMatchingRoot) return false
+            if (!super.equals(other)) return false
+            return true
+        }
     }
 }
 
@@ -156,7 +236,8 @@ class CertificateInvalidException(message: String, cause: Throwable? = null, val
  * @param cause the underlying exception
  * @param reason one of a set of well-defined [Reason]s why the attestation failed
  */
-class RevocationException(message: String?, cause: Throwable? = null, val reason: Reason) : AndroidAttestationException(message, cause) {
+sealed class RevocationException(message: String?, cause: Throwable? = null, val reason: Reason) :
+    AndroidAttestationException(message, cause) {
 
     /**
      * Possible reasons, a [RevocationException] was thrown
@@ -188,5 +269,42 @@ class RevocationException(message: String?, cause: Throwable? = null, val reason
         var result = super.hashCode()
         result = 31 * result + reason.hashCode()
         return result
+    }
+
+    class ListUnavailable(message: String?, cause: Throwable? = null) :
+        RevocationException(message, cause, Reason.LIST_UNAVAILABLE) {
+        override fun equals(other: Any?): Boolean {
+            if (this === other) return true
+            if (other !is ListUnavailable) return false
+            if (!super.equals(other)) return false
+            return true
+        }
+
+    }
+
+    class Revoked(
+        message: String?,
+        cause: Throwable? = null,
+        val certificateChain: List<X509Certificate>,
+        val revokedCertificate: X509Certificate
+    ) : RevocationException(message, cause, Reason.REVOKED) {
+
+        override fun equals(other: Any?): Boolean {
+            if (this === other) return true
+            if (other !is Revoked) return false
+            if (!super.equals(other)) return false
+
+            if (certificateChain != other.certificateChain) return false
+            if (revokedCertificate != other.revokedCertificate) return false
+
+            return true
+        }
+
+        override fun hashCode(): Int {
+            var result = super.hashCode()
+            result = 31 * result + certificateChain.hashCode()
+            result = 31 * result + revokedCertificate.hashCode()
+            return result
+        }
     }
 }