+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## NEXT
 * Attach more context to exceptions
+* Verify that patch levels are not too far in the future (default leeway: 1 month)
 
 ## 1.8.2
 * Integrate Google's new PKI cert path validator